application/requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
nose==1.3.7
validator-collection==1.3.5
PyMySQL==0.9.3
Flask-Migrate==2.5.2
